* configure.in config/amanda/progs.m4 config/amanda/flags.m4:   use AC_PROG_EGREP instead of the Amanda version * server-src/find.c: fix compiler warning on Solaris 8
git-svn-id: https://amanda.svn.sourceforge.net/svnroot/amanda/amanda/trunk@804 a8d146d6-cc15-0410-8900-af154a0219e0

+++ server-src/find.c
@@ -264,7 +264,7 @@ find_compare(
 
     for(k=0;k<nb_compare;k++) {
         char sort_key = find_sort_order[k];
-        if (isupper(sort_key)) {
+        if (isupper((int)sort_key)) {
             /* swap */
             sort_key = tolower(sort_key);
             j = *(find_result_t **)i1;
